requirements.txt
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
rich==13.7.0
schedule==1.2.0
slack-sdk==3.26.1 